+++ packages/ui/package.json
@@ -16,16 +16,17 @@
   },
   "dependencies": {
     "@origin-space/image-cropper": "^0.1.9",
+    "@tanstack/react-table": "^8.21.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.1.1",
     "lucide-react": "^0.478.0",
     "next-themes": "^0.4.4",
     "radix-ui": "^1.4.2",
+    "react-dropzone": "^14.3.8",
     "react-hook-form": "^7.54.2",
     "sonner": "^2.0.1",
-    "tailwind-merge": "^3.0.2",
-    "@tanstack/react-table": "^8.21.2"
+    "tailwind-merge": "^3.0.2"
   },
   "devDependencies": {
     "@agentset/eslint-config": "workspace:*",
