+++ packages/ui/tsconfig.json
@@ -4,10 +4,6 @@
   "compilerOptions": {
     "lib": ["ES2022", "dom", "dom.iterable"],
     "jsx": "preserve",
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    },
     "module": "esnext"
   },
   "exclude": ["node_modules"]
